requirements.txt
# Core Data Science Libraries
pandas>=1.5.0
numpy>=1.24.0
scipy>=1.10.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=1.7.0
lightgbm>=3.3.0
tensorflow>=2.13.0
keras>=2.13.0
imbalanced-learn>=0.11.0

# Model Interpretation
shap>=0.42.0
lime>=0.2.0
eli5>=0.13.0

# Data Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
bokeh>=3.2.0

# Dashboard and Web Interface
streamlit>=1.25.0
streamlit-aggrid>=0.3.4
streamlit-plotly-events>=0.0.6

# Statistical Analysis
statsmodels>=0.14.0
pingouin>=0.5.3

# Data Processing and Feature Engineering
category-encoders>=2.6.0
feature-engine>=1.6.0
optuna>=3.3.0

# Time Series Analysis
prophet>=1.1.4
pmdarima>=2.0.3

# Anomaly Detection
pyod>=1.1.0
isolation-forest>=0.1.0

# Database and Data Storage
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pymongo>=4.4.0

# Configuration and Environment
pyyaml>=6.0
python-dotenv>=1.0.0
hydra-core>=1.3.0

# Logging and Monitoring
loguru>=0.7.0
mlflow>=2.5.0
wandb>=0.15.0

# Testing and Quality
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
black>=23.7.0
flake8>=6.0.0
isort>=5.12.0

# Jupyter and Documentation
jupyter>=1.0.0
jupyterlab>=4.0.0
notebook>=7.0.0
nbconvert>=7.7.0

# Utilities
tqdm>=4.65.0
joblib>=1.3.0
requests>=2.31.0
click>=8.1.0

# Development Tools
pre-commit>=3.3.0
mypy>=1.5.0

# Performance
numba>=0.57.0
dask>=2023.7.0

# Geospatial (for location-based fraud detection)
geopandas>=0.13.0
folium>=0.14.0

# Text Processing (for transaction descriptions)
nltk>=3.8.0
spacy>=3.6.0

# API and Deployment
fastapi>=0.101.0
uvicorn>=0.23.0
gunicorn>=21.2.0

# Cloud Services (optional)
boto3>=1.28.0
google-cloud-storage>=2.10.0
azure-storage-blob>=12.17.0
